Move segyio/_trace.py -> segyio/trace.py
In preparation for a better, more well-documentated and cleaner
implementation of the trace mode, move it to not be _-prefixed.
