+++ python/segyio/trace.py
